labstore: long-overdue refactor of primary profile and role
This moves the parameters for the primary nfs cluster into hiera.
The shared values that will be needed to manage the backups better are
in the common/ namespace vs the role hierarchy so that the backup servers
can figure out which server to connect to. This should allow us to
do a longer-term failover for replacing servers, maintenance, etc. after
a refactor of the backup profiles.

This introduces a custom fact that is a list of drbd volumes with the
role (not the state) of the volume. That way things that are supposed
to run on the drbd primary of the cluster can be triggered only when
all volumes are actually in the primary role. It will also make it clear
when the actual role isn't the expected role.

Overall, this should allow more re-use of this code in the near future
for the secondary NFS cluster.

Also added some comments where there are obviously patches needed to
clean up or fix things not addressed immediately in this patch.

The fact has been tested on labstore1005 with
'facter --custom-dir'.

Change-Id: If0f28e7bf38979a70c3e6fd14350b5067d763713