set disable_fsnotify for all current mtail usage
* clean up 3.0.0~rc35 transition settings
* add pin to mtail 3.0.0~rc35-3+wmf1 and enforce because -disable_fsnotify is not a flag
in versions prior to 3.0.0~rc22.  enforcement is to try to prevent long-term breakage
in other environments.

NOTE: the upgrade path for mtail to 3.0.0~rc35 is to install the package in one step and
reset the systemd unit in another.  there is no explicit dependency in puppet so I foresee
the possibility of needing a second puppet run to complete the upgrade and ensure mtail
is running.

Bug: T251466
Change-Id: I0fcaca3e93b9ec35f96c502cba14beb7fcb45664