nfs monitoring: fix the broken paths for the directory size monitor
When we removed bind mounts, the /exp paths became meaningless.
If we are failing over, this cron could be a blocking file lock, so it
needs to be killed by the failover script in case it is running.

This also removes some old comments about the bad old days and adds
a friendly "done" message to the end of the important ops in the
nfs-manage script.

Bug: T160113
Change-Id: I0bc062388757c6cb4c29ccfcf2d81463c87fa7f4