wmcs: kubeadm: decouple haproxy profile for Toolforge/PAWS
Before this patch, the haproxy configuration for Toolforge and PAWS was coupled, it was very
similar. But as we keep developing the new PAWS setup, the config is forking more and more.

Decouple the profiles now for smoother future development.

While at it, do some small adjustements to both hiera and the templates to accomodate to this new
layout.

Bug: T195217
Change-Id: Iaf84c6805d54fb0b79543c80646458c7263489c2
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>