gerrit: Quote `container.javaOptions` in config
Ever since we added flogger configuration to the Gerrit's java
options, the value contained hash mark (#), which git config treats as
comment character, if unquoted. Hence, gerrit treated everything past
the hash mark as comment, and failed to detect that a flogger config
is in place. So upgrading Gerrit added the Flogger configuration once
more.

While our start-up script does not rely on `javaOptions` in
gerrit.config, we nevertheless do not want to have misleading values
there. So we quote the string to properly treat hash marks.

Change-Id: Ic429c8ea30614f08933fb47be8a9a5ac63f94ea3