mediawiki::maintenance: add server-header config
What: Ensure mw-maintenance servers return their FQDN in
      response headers just like mw-appservers already do.

Why:  Users debugging can't tell which server they are talking
      to and just get "Apache" for mwmaint* unlike other mw* servers.

root cause:

This puppet code is used in profile::mediawiki::httpd which in turn
is included in profile::mediawiki::webserver which is used in:

role::mediawiki::appserver
role::mediawiki::appserver::api
role::mediawiki::parsoid::testing

but NOT in role::mediawiki::maintenance, the role for mwmaint servers.

Since the full webserver is not needed there, the fix above is to just
add the server_header snippet.

Bug: T255629
Change-Id: I6a682e14384fba1ba9e6b6297293a5d899d12418