unattendedupgrades: allow configurable kernel cleanup
People have wanted to be able to clean up unused kernels automatically
at times. I'm sure this is safe in some cases, however, I don't think
it is considered best practice to do automatically across the board
since some VMs are not rebooted often and kernels are tricky at times.

This will keep it off by default, but we should be able to enable it
where appropriate or desired. It should be easy enough to test one VM
at a time.

I also removed the :: at the end of the cleanup interval key because
appending seven days is not what I want to do in any case :)

Bug: T127374
Change-Id: Icb54dbca33aec1d2721bd30513c04a7cc3df553d