zuul: add gerrit-test:29418 as a ssh known host
The production Zuul listens for gerrit-test.wikimedia.org which has the
same ssh host key as gerrit.wikimedia.org. But given they have different
hostname/IP, the key is not recognized as a known host.

Readd the same key with gerrit-test.wikimedia.org as the hostname.

Hosts: P:zuul::merger

Bug: T253263
Change-Id: I60e595f1dcc99e89a29bc3cfe9d91578323fc970