phabricator: Hiera'ize ferm srange for port 80
In cloud we need to allow port 80 connections from proxy
servers because they don't speak TLS to backends.

In prod we don't need to open it because ATS speaks TLS
to backends and TLS terminates with envoy on the phab host itself.

In both cloud and prod we want to open port 80 for deployment
servers for testing.

Change-Id: I40c203e5d09dd40e33034bdc7a46c603546c9a9d