dnsdist: add parameter for web server configuration
Add a parameter to enable dnsdist's web server configuration. This also
enables the API endpoint for /metrics, which we use for sending
statistics to Prometheus. Also update the wikidough profile as well to
reflect the change in instantiating the dnsdist class.

Additionally, update the keys in the wikidough.yaml file by prefixing
the correct class to which they belong.

Bug: T252132
Change-Id: I1af5fb94f343818e9c0909e8a05933f795a1a0d2