gerrit: Add option to mark gerrit servers as upgraded
The config files that we use for our current Gerrit 2.15 don't work
for the upcoming Gerrit 3.2. Since different Gerrit servers will run
different Gerrit versions for some time, we add a parameter to
distinguish which server is running the current/old version and which
has already been migrated to the new one.

Bug: T254158
Change-Id: I7f4aaca47695d690517d2f0ae0248dc4fd437375