haproxy: add some settings for tcp backends
These are tuned for mysql, the only thing we're using
the tcp template for currently. If someone wants
to use this backend for a different purpose some
of this needs to be made configurable.

OpenStack keeps long-lived connections open to mysql
and gets upset if they're closed by someone else.  These
settings allow us to keep those connections alive
for as long as OpenStack is expecting (60m)

Bug: T242455
Change-Id: I337402fc048d87e2363b94c94a795897789cce1a