contint: remove obsolete firewall rules from labs
We have ferm rules for the labs instances to allow the Jenkins master to
ssh into them. However they are not applied and can be dropped.

That is a legacy from when we used Puppet to provide the test
environment.  Some classes included ferm and their rules which in turn
caused to default to drop everything.  This contint::firewall::labs
class was to punch holes in the firewall.

Fast forward to 2020, we no more use Puppet to provide the test
environment. ferm is never included and the config files provided by
ferm::services are never realized since the ferm class is not included.
The reason is that the config files are virtual resources (@file) and
ferm::init is required to have them created (it has: File <| tag ==
'ferm' >).

Thus drop the now unused class contint::firewall::labs.

That also resolves a wmf-style issue :)

modules/profile/manifests/ci/slave/labs/common.pp:10 wmf-style:
profile 'profile::ci::slave::labs::common' includes non-profile class
contint::firewall::labs

Bug: T114209
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: Ic1babd23d08818274fccd68c7b7a061831d11708