gerrit: Allow to use request tracing for new Gerrits
The current production Gerrit does not use Flogger, and hence the
change is a noop there . Gerrit 3.x uses Flogger and setting the
logging context allows to enable request tracing. This should help if
users run into issues, as it allows to add a tag to the request and
log this tag throughout the request processing within Gerrit. This
allows to reproduce which path the requests take in Gerrit and should
make bugs reproducable.

Bug: T254158
Change-Id: I4220046509a0ca60932f21da5621679727f8d6ee