profile::mediawiki::mcrouter_wancache: send probe after 60s
With the actual config mcrouter waits only 3s before sending health probes
to a shard marked with TKO, to verify if it is reachable again or not.
This change increases the time to wait to avoid unnecessary back and forth
from the TKO state for one or more shard during a high traffic event.
The value is kept reasonably short to allow speedy recoveries when needed.

Bug: T255511
Change-Id: Ibd30517f61865ef9c37a66b92dac80a39aae6c3e