ci: switch integration.wikimedia.org to scap DocumentRoot
The git checkout at /srv is left behind, it is polluted by entirely
different files that we need to keep. That will have to be manually
cleaned up.

We tried this change previously but had to rollback to an issue in the
PHP source code. It was not supporting the scap managed document root
being a symbolic link. That has now been fixed and deployed.

Hosts: contint2001.wikimedia.org

Bug: T149924
Depends-On: I369ce9a26d89f52bb90f5fc1eccc47804a6936e0
Change-Id: I499274a0bafe0d1699422be10edaad7af51efd4e