wikidough: update dnsdist web server listen address
Currently, dnsdist listens on localhost for its web server. This web
server is also used for the API and the /metrics endpoint to export
statistics to Prometheus. To allow this, listen on 0.0.0.0 instead of
localhost and additionally configure the firewall to restrict access to
traffic from $PRODUCTION_NETWORKS.

(Note that dnsdist 1.5.0 provides an "acl" parameter that allows an ACL
netmask list to be specified in the webserver() function but we can't
use that since we are currently using dnsdist 1.4.0.)

Bug: T252132
Change-Id: Id608a9d907e1866f4be6e4fa585d9951cd165f7e