toolforge: mailrelay: enforce ratelimiting
Previous to this patch, ratelimiting was only set to "warn", but it wasn't enforced.
Turn enforcement on.

While at it, put the log_message statement last, so it can correctly read the $sender_rate_limit
variable (which is set by the ratelimit statement).

Also, fix the syntax of the ratelimit config file to include just one asterisk, otherwise the
search doesn't match and exim rejects all the email (wrong ratelimit is configured).

Bug: T175964
Change-Id: I918bf54f45e752b7c71b9719715b31b091f06563
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>