lvs: Add new proton TLS service
Why:
We want to move over proton from the old infrastructure to the
kubernetes one

What:
Add a new LVS service. Same IP, different port than proton
Start with state: service_setup
Rename the old proton service to denote that it is not TLS enabled.
Amend the description as well to showcase that

Bug: T225680
Change-Id: I04f53100e0bdc385e9708223c6066c8a6eb2ddd4