profile::reportupdater::jobs: absent old RU job
It seems that the RU page-creation job still queries the log database
on db1108, but we stopped pushing data to it months ago so reports
have been showing a steady 0 during this time.
This change absent the job to avoid queries (db1108 will be reimaged
soon and repurposed), we'll decide later on if we need to move the
job to Hive or not.

Also remove profile::reportupdater::jobs::mysql since it is not used
anymore and it causes confusion.

Bug: T234826
Change-Id: I4c8dc5d755087a3353c1d580d5575a037d0d0ea7