prometheus: use the correct password for the wikidough job
In ab8a948a, we added the Wikidough job to Prometheus. However, the
metrics were not being fetched from the malmok host and on debugging we
found that the issue was that for the basic_auth password, we had to use
the web server password and not the web server API key.

The reason for this confusion stemmed from the fact that to enable the
/metrics endpoint, you need to enable the API and that is done by
setting the API key in the webserver() function. That works if you set
the "X-API-Key" header (which we can't do through Prometheus in any
case), but for basic_auth you still need to use the web server password
and not the API key.

Bug: T252132
Change-Id: Iac865d6118cdfbe2cd7221e2dce76db26a5a2512