toolforge: mailrelay: introduce SRS to correctly envelope forwarded emails
This patch introduces SRS into toolforge/toolsbeta Exim. The config/code is adapted from this URL:
 https://github.com/hn/exim-misc/blob/master/exim-srs-sender-rewriting-scheme.conf

The problem we are trying to solve is related to mail security in the Toolforge forwarding
scenario. If we don't correctly envelope forwarded emails, the final destination smtp server may
reject emails as spam because the SPF checks wont pass, given Toolforge email relay is not a valid
sender for the original email From address.

Bug: T120225
Change-Id: If0c65ea35ce50aa501e2e9b8ed290bec7df251bd
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>