releases::mediawiki: remove PHP packages
The hosts run Jenkins and an idea was to have them cut the MediaWiki
release tarball automatically then run all the tests with the resulting
tarball. The idea required PHP.  That was back in January 2018 but never
concretized in anything useful.

None of the deployed Jenkins job rely on PHP at all and most should be
deleted anyway (T256164).

Remove composer as well (T249949).

Bug: T249949
Bug: T256164
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: If0056a0a277d41f7ab960f2d9ff50f2fb09e884e