zuul: set site/initsystem in rspec configuration
initsystem comes from our base module and is not set by
rspec-puppet-facts. That thus lead to an error whenever we include a
module that uses it such as systemd.

Set it globally in the spec_helper.rb

For the same reason, set the site to always be eqiad. It is thus no more
needed to set site as node parameters which was a bit of a hack to set
the facts.

Change-Id: Ib53b9ea3e989f3a110e45a423c7c95ec142d1ad5