gerrit: As log4j.xml is a static file, treat it as static file
The Log4J config was once a template when we fed directly to
logstash. As that part is gone since
592e76e97c60c71e9c3fdf29b39982a7022f0b55 .

As it is a static file, and we'd need to tune it a bit for Gerrit
v3.2, we move it (unchanged) to the homedirs, and add the Gerrit v3.2
variant in follow-up commits.

Change-Id: I5b153d0ffb172456b63ddd2f3e6ed7d20d508d7a