mediawiki: Change mw alerts to use a moving average
This used irate() which compares the two newest two data points from
the 10min range and extrapolates from it the current estimated rate.

This is useful for a graph, where sub-minute precision can be valuable. [1]

However for an alert, I think it makes this overly sensitive. I don't think
we should act differently between 60 fatals happening in a 5 second span,
with nothing the rest of the minute, or continuously about 1 every second.

With rate() you get the average over the given range, with irate() you get
the delta between the last two counter positions which could be as little
as 10 or 15 seconds apart in production.

This has the obvious downside that if there were zero fatals for a period of
time and then enough to be more than our threshold, it will take a few
minutes for the average to reach the threshold. As such I've shortened
the time span from 10min to 4min.

As far as I know, this is mostly a catch-all alert as fallback for other
alerts. There are healthchecks and other dedicated alerts which have
strict boolean assertins that report much quicker. For the unknowns that
we don't monitor, we have this.

[1] https://www.robustperception.io/irate-graphs-are-better-graphs

Change-Id: I74fa06a92fa9883d1a4d3629eda5446f1f68d884