mediawiki: Remove 'test.wikimedia.beta.wmflabs.org' vhost
This wiki is configured as a Wikipedia production at test.wikipedia.org,
and is (mostly) configured that way in Beta as well.

Now, because of the weird and wonderful magic that is Multiversion, it is
in fact already accessible both through https://test.wikipedia.beta.wmflabs.org
and https://test.wikimedia.beta.wmflabs.org.

The former works because *.wikipedia is mapped.

The latter works because this vhost is explicitly created Puppet
and explicitly added to Multiversion in Beta.

Bug: T99156
Depends-On: I56eb4a802940af1992ad9d33e392a1f9db6721aa
Change-Id: I135f54d7d27997922c67b11b560fa44197f0e2f4