zuul: stop prefixing report with the job name
We had the Zuul build report to be prefixed with the raw job name to
simplify pretty formatting in Gerrit using a commentlink hack. With
Gerrit 3 that is no more possible, the prefixed job name is shown which
is rather noisy:

* test-job https://www.example.org/jenkins/test-job : SUCCESS in 42 s

Remove the prefix to just show the build URL and status. The report
would thus show as:

* https://www.example.org/jenkins/test-job : SUCCESS in 42 s

Bug: T256575
Change-Id: Ie789ca9dda6226ac55f481d0c1a2e1a7cba463a3