gerrit: Move new version's homedir to default place
Since the new Gerrit version is the standard version now, also the
homedir should be in the standard place. So we move `homedir-new` to
`homedir`.

(If we merged this change into the commit that removes the
`is_new_version` parameter, the files in homedir would show diffs,
which is not what we want. We want to delete the homedir files, and
move the homedir-new files onto their place. Hence, the whole process
it two commits)

Change-Id: I7682bb037359a2089403d31f57f166b87df737dd