webperf: Serve different robots.txt on beta site
We don't want search engines to index the content on
performance.wikimedia.beta.wmflabs.org, and then decide that the
copy on performance.wikimedia.org is a duplicate.

This aliases /robots.txt in the Apache configuration to stop the
non-canonical copies of the site from being crawled.

While I was at it, I fixed a naming inconsistency with the Apache
configuration template.

Other than a new file in /var/www, this is a no-op in production;
see: https://puppet-compiler.wmflabs.org/compiler1003/23623/

Bug: T255092
Change-Id: If76f308f0b48ee7498bbcb0c9e44d8f1d45eb366