mariadb-backups: Fix remote_backup_mariadb.py deps and path
When modifying remote_backup_mariadb.py, python3-yaml dependency was
lost, this was not an issue because cumin also depends on it, but
it wasn't right.

The main issue was that transfer.py was called with absolute path,
and now it changed location, so all snapshots failed execution.
Fix this by calling transfer.py with no absolute path.

Bug: T138562
Change-Id: Ibad032062b48ba11bc5f96c596acce7ac837a877