Add discovery records for chartmuseum
Add a new chartmuseum service in service::catalog. Specify most of the
stanzas, but skip specifically the lvs stanza.

In order to allow for that, make that stanza Optional in wmflib::service
type and check for existence in Wmflib::service::get_services_for_lvs
Also amend the wmflib::service::fetch function to allow returns only
services that specify an lvs stanza

Also add a notes_url Optional stanza in wmflib::service::monitoring, to
allow services to not default to the usual LVS icinga notes_url

Then use it for chartmuseum, putting a comment to denote that

Change-Id: I876a03497f2e6a63dd18b1d9a65f4f0e71b10698