mediawiki: Create /etc/firejail/mediawiki.local
This enables most of the restrictions already present in
mediawiki-converters.profile to be found by the default MediaWiki shell
wrapper (FirejailCommand).

It has the extra addition of blacklisting /srv/mediawiki/private as an
extra hardening measure.

mediawiki-converters.profile is left in place (it includes the new
mediawiki.local) as some commands are still manually using it.

Change-Id: I14e97e47d65f50ce4ba928d5857607fdaf9d478a