gerrit: Add table for Zuul CI results underneath the commit message
This module is from

  https://github.com/dburm/pg-test-result-plugin

under the MIT license.

This change brings much information to the table. We (for now) keep
this information to expose what is possible, and remove clutter only
in follow up commits. That way, people can always come back to this
commit to see how to for example add the name of the checker back in,
if we ever choose to.

Bug: T256575
Bug: T206453
Change-Id: Ib8b77741b684ac8b4f4c2e5b2e42ad4bfa9d70b6