zuul: remove gerrit-test connection and setup
This is to start removing the gerrit-test setup again
that had been created to prepare the Gerrit 3.2 upgrade.

Remove the additional connection and setup in zuul.

Unless there is an opinion we should permanently keep
a test setup in production like this.

Though the original request said it would be reclaimed
after the migration is complete.

Bug: T239151
Change-Id: I54cbc6ee9c7b73ef122f4982550a44ceb72d82cf