varnish: Facebook temporary experiment is permanent
The few things we have tried to reduce excessive traffic from Facebook
have worked well. Remove the special rate limiting threshold for fb now
that with 7b9d83b we have lowered the limits to 100 rps for everyone.

Update the comments to reflect the fact that these countermeasures are
in fact not temporary.

Bug: T192688
Change-Id: Iaa817987309b817f0a40e407a2777f6dba24e0cf