ATS: handle backend checks at healthcheck.wm.org/ats-be too
To make healthchecks a bit less confusing as part of T255015, we have
decided to consolidate on one single hostname for the Host header:
healthcheck.wikimedia.org. Also, the path component should be named
after the intended recipient of the heathcheck instead of the sender or
something else entirely.

Make sure ats-be handles checks sent by varnish-fe to the URL
http://healthcheck.wikimedia.org/ats-be. After this change is rolled
out, we can update the actual URL used by the frontends themselves.

Bug: T255015
Change-Id: I87b22d1f5ae8ab2e339520ea313ba98db6853727