Don't include "backports" and "thirdparty" components into Stretch images
Starting with stretch we stopped using the backports component (which was
folded into "main", not to be confused with backports.debian.org) and the
"thirdparty" component (which was split into thirdparty/*) on
apt.wikimedia.org. The components are still present on apt.wikimedia.org
to not break existing systems polling it, but "backports" is entirely
empty and "thirdparty" only has a mis-imported oath package.

So remove them from the base image, no need to add further references to
the unused repos.

Bug: T257327
Change-Id: Ic2b5bfb122ad9d0fc7f4e404f639d9b71114691f