tools-prometheus: set up prometheus to get paws metrics
Since we are still using tools-prometheus for all VM metrics and trend-type
observability, this will allow us to gather the same for the paws project
cluster. The job names and namespaces should allow us to prevent pollution
of the toolforge metrics on our dashboards (with a little adjustment).

That said, this will certainly merit a bit of review!

Bug: T256361
Change-Id: Ifb6b4811fea89a9cbe4fbb0f414ac07313d30c69