tools-prometheus: set up prometheus to get paws metrics
Since we are still using tools-prometheus for all VM metrics and trend-type
observability, this will allow us to gather the same for the paws project
cluster. The job names and namespaces should allow us to prevent pollution
of the toolforge metrics on our dashboards (with a little adjustment).

This version should have fixed the jupyterhub target.

Bug: T256361
Change-Id: I2ee00d04c706dda3a910e76dd20299fd41b0ca09