toolforge: enable delete API for docker-registry
WE have a bunch of old cruft that we would like to clean out of the
docker-registry at https://docker-registry.tools.wmflabs.org. To allow
this cleanup we need to enable the manifest and layer deletion API for
this deployment.

See <https://github.com/docker/distribution/commit/9c1dd69> for the
configuration needed.

Change-Id: I4db11a1b4e44851c6b6a12fb5898be4b9734e59f