releases::mediawiki: support rsyncing files to multiple secondaries
We want to be able to rsync releases files from one source to
multiple destination servers to be able to setup new releases
backends on buster.

This first does it for MediaWiki and then other files will follow
and eventually we can remove the single "passive_server" in favor
of an array of multiple "secondary_servers".

Bug: T247652
Change-Id: I371e6fd306891576ebbb87e1972a7e410990e6a3