Explicitly mentions the repository in scap::sources
scap::sources defaults to prefix the resource title with
'mediawiki/services' to forge the Gerrit repository name. It made sense
when MediaWiki services were the first adopters of scap, but the scope
is way larger now and the default is not so intuitive anymore.

Explicitly set the repository parameter for each source that ommited it.

While at it fix docker-pkg/deploy using a double quoted key.

Hosts: O:deployment_server
Bug: T257413
Change-Id: Ibcc70cce208c6faee87ce1ccc67a68dd3ea9a7a8