toolforge: urlproxy: drop support for the legacy routing scheme
The toolforge front proxy no longer receives traffic to the tools.wmflabs.org domain, and therefore
we no longer support the legacy routing scheme.

We can simplify the setup by assuming everything is using the toolforge.org domain and, if not, the
tool-fourohfour webservice running in k8s will handle it.

Bug: T234617
Change-Id: I9de66b9e88977095085ebbac934cb6ae5dda7c4c
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>