rsync::quickdatacopy: add optional parameter to let rsync --delete files
Add an optional parameter to add rsync's --delete parameter.

This will delete files on the _destination_ side that do not exist
(anymore) on the source side.

The use case here is a typical rsync of data between multiple servers
like MediaWiki release files synced between releases1001 to releases2001.

Without the --delete parameter files that are deleted on the source
will pile up on the destination sides and never be removed unless someone
manually removes them and the disk space on destinations will continue
to shrink.

Instead we want exact mirrors where both sides are always the same size in these
cases.

Bug: T247652
Change-Id: I8a94a7dff308ee270bbe0ed09cca501708cd7336