releases: remove duplicate rsync code from blubber and parsoid classes
The rsync in the mediawiki class already syncs the entire
/srv/org/wikimedia/releases directory.

As subdirs of that blubber and parsoid are already included
and repeating the rsync is just duplicate.

Also removing the passive_server parameter in the mediawiki
class which is not used anymore in favor of secondary_servers
which allows for multiple destination servers to sync to.

Change-Id: I269773709a52a15c8c3ee9481ba33b0f079c28dc