paws-prometheus: add node exporter info to tools-prometheus for paws
Many of our dashboards assume we have node exporter stats as well as
k8s stuff, specifically for the k8s control plane and etcd (which in
paws are the same thing). This will add the paws project stats.

If we only want the control plane, the last patchset will do that.
Providing both for argument's sake.

Bug: T256361
Change-Id: I8cfab95dc2d5e8cdedb5b6c50af9427a81b2f536