tools-prometheus: reduce scrapes and increase scrape_timeout on cadvisor
On very busy Toolforge nodes cadvisor takes 44+ seconds to respond.
This is why our prometheus is showing so many cadvisor pods as "down".

I am proposing scraping every 4 minutes and increasing the timeout to
60s so that we get stats from those pods. I am not sure which dashboards,
if any, these stats feed, but I hope they can cope with that. They are
missing stats as is.

Change-Id: I1c7799279aa4fe5d2d4055181b38edf0eb8b3fbe