kubeadm: If using a stacked control plane, expose etcd metrics
By default the metrics URL is only availble to localhost on the control
plane node. This is not helpful for us to gather prometheus stats with.

Tested manually on bstorm-testk8s-control-2.testlabs.eqiad.wmflabs

It can be retro-fitted to the paws cluster by editing the manifests
and updating the kubeadm config map.

Bug: T256361
Change-Id: Id30b500f9bf591cc389ce64a8ff7048bed456530