bacula: Add ignorelist for long-term broken backups
In an ideal world, backups should be configured for services that
actually need them and can produce them. In reality, hardware can
fail for a long time or host/services can be in an unavailable
state for a long time.

This is why an ignore list functionality has been added for both
prometheus and icinga monitoring, so no alerts happen for disabled
jobs, as configured manually thought the job_monitoring_ignorelist
file on puppet.

Now, when check_bacula.py is run with either the --prometheus or
--icinga options, jobs listed on that file are skipped. This does
not apply for command line arguments such as manually running
check_bacula.py, checking an individual job or listing the
configured jobs.

Backups are still retried (they are not disabled), this only
modifies monitoring. In a way, this is implemented as there isn't
a good way to partially ack a service check on icinga.

The ignore list allows comments, which point to the task that
should be resolved before jobs being removed from it.

Bug: T234900
Bug: T256846
Bug: T244530
Change-Id: If6a3ad47824101bd427ce66bd8f3d7431c6ba2c5