Remove apt pin for stretch-backports for npm
Stretch-backports is going away now that Stretch is in LTS support.

Drop the apt pin for npm, restreduce will be moved to Buster and the
current version from stretch-backports is incompatible with our
node10 component (clashing package relationships), so we also can't
simply mirror it internally.

The existing packages on scandium will remain installed until it's
eventually reimaged.

Bug: T256877
Bug: T257906
Change-Id: I452f7fda8ec57e0a47a371dbb737f219a7fff617