cloud-nfs: Allow changing the nfs mount version
This need to be a hiera setting because for now, the version of systemd
and NFS on stretch bastions don't seem to play nice together with any NFS
version over 4.0.  While I'd like to figure that out, we need to ensure
the NFS version stays the same on bastions if we upgrade to 4.2 elsewhere
in Toolforge.

Bug: T257945
Change-Id: I03dfbb843bae1d6561f9380bf3dcdee0c550b8bc