toolforge: Perform HTTPS redirects unconditionally
Close the "POST loophole" that was allowing POST requests to bypass TLS
security by changing the HTTP -> HTTPS redirect logic. The new logic
makes a separate server block only for handling HTTP requests and
unconditionally redirecting them to the matching HTTPS endpoint. The
status code used is 301 rather than the more technically correct 308 out
of an abundance of caution with dealing with legacy clients that have
not be updated for RFC 7538 (2015).

Bug: T102367
Change-Id: I13fea55e2bd622ee3b0c9a88db32a6604a5f5404