mariadb-backups: Modularize common stats config into a subtemplate
As the number of configurations of the different backup jobs grew,
it was apparent that some of the configuration was repeated
throughout all jobs. To make at a later time easier to change
where the metadata for the backup jobs live, remove that config
from every backup configuration file on puppet and refer it on
a subtemplate, so it can be changed once and apply everywhere.

Bug: T258045
Change-Id: Ie933aadd1eb4e7b2a26d15a36152b6750fb25046