dnsdist: reload the certificates instead of restarting the service
dnsdist does not support reloading the configuration (see
https://github.com/PowerDNS/pdns/issues/6249) but it does support
reloading of certificates through the 'reloadAllCertificates()' command.
We should restrict restarting the dnsdist service through Puppet as it
is not transparent, and clears the cache, among other issues.

This commit removes the notification to restart the service when the
configuration file is updated and instead changes "restart" to mean
"reload", as there seems to be no way of doing a "reload" through Puppet
directly.

To handle the reloading of certificates (managed by acme_chief), we add
an "ExecReload" to the dnsdist service instead, for which an override is
added. This means that while certificate updates will be handled
automatically (through reload), restarting the dnsdist service in case
there is a configuration change will have to be performed manually,
which is an approach we prefer.

Hosts: malmok.wikimedia.org

Bug: T252132
Change-Id: I52fb7ac249e7d5f49f6b7ed129e6462da054968e