parsoid: create new role to install just testreduce
Create a new role that install just the testreduce visual
diffing server / client part of the existing parsoid-testing
role.

This should pull in nodejs and npm on buster.

testreduce is written in Nodejas as opposed to current
PHP Parsoid.

Also add the parsoid admin groups on the new roles,
same as they are on the existing parsoid::testing role
which is applied on scandium.

Finally apply the role on the new VM testreduce1001.

Almost certainly the role will need follow-ups, this
is to get it started, install nodejs and npm and let
Subbu have access.

Bug: T257940
Change-Id: I7ea16a1038c87ed29972ff6272e7e80c0385313b