visualdiff: update git branch from ruthenium to scandium
The branch "ruthenium" does not exist anymore. Therefore
the puppet git pull command on a first run fails with:

fatal: Remote branch ruthenium not found in upstream origin

https://gerrit.wikimedia.org/r/admin/repos/integration/visualdiff,branches

shows that there is the scandium branch though and scandium
replaced ruthenium.

Set this to scandium for testreduce1001 and scandium itself.

Though:

[scandium:/srv/visualdiff] $ git status
On branch ruthenium

Bug: T257906
Change-Id: Ida6cc2427ef9603625d3dee11987cefa3094dff4