profile::hive::client: allow to specify jdbc host:port
The hive client profile is a shared puppet config to create
the hive-site.xml config file, hence used by hive daemons and their
clients. If the Analytics team needs to move the hive daemons to
another node (for example, during a failover), there should be an easy
place in hiera where to tune these values.

Bug: T257412
Change-Id: I8c5b6667989b3a1b4b44485234c0e5d592c62e51