dnsdist: add a parameter to set the size of the ring buffers
dnsdist uses ring buffers for inspecting live traffic (for functions
like topClients(), topQueries(), etc.) and the size defaults to 10000.
This commit adds a parameter to the dnsdist module to configure the size
and sets a default value of 10.

The reason for this change is to limit the amount of data we collect.
For example, restricting the size of setRingBuffersSize() to 10 will
restrict topClients() to only the (top) 10 clients, instead of 10000
clients, which is dnsdist's default.

Bug: T252132
Change-Id: Ic25f295a6bdce4b4974fa0c88d4d438bb954fb98