Add mw_resource_loader_uri to Node.js service config vars
Adds a new mw_resource_loader_uri config variable for use in templated
requests to load.php. Mobileapps was previously making raw requests to
the public load.php endpoint, but this does not work in Kubernetes.

This is needed to support the mobileapps service as it transitions off
SCB and onto Kubernetes.

Bug: T258186
Change-Id: Ia436ac9249170d27e4d1dece3eebd8efae86b64c