mariadb-backups: Adjust check parameters to get less false positives
A 1% variation on database size from one week to another is not
that rare, given that we are having a 4% increase in the image
table on commons currently. It is also over the growth of the
read-write es hosts.

Set the default of critical to 15%, which was what was actually
used on input parameter, and align all defaults to prevent
reading confusion.

Additionally, make the minimum backup size 300K, as it is now
the size of the zarcillo backups after dbbackups metadata has
been moved to m1.

Bug: T258045
Bug: T138562
Change-Id: I78918fad744577ad4a171f7f8a588dcc3ad007e9