otrs: Remove demime condition from exim
Why:
It's a deprecated condition for more than 10 years and has been removed
in exim 4.88. Buster comes with 4.92 and no longer recognizes it as a
directive

What:
Remove demime condition from the acl_check_data ACL. There does not seem
to be a reason to replicate somehow the functionality now. OTRS exim is
not accessible to the public internet and the MX servers that are placed
in front of it anyway have protections as well. clamd malware scanning
will continue working anyway, it does not require the demime action

Bug: T187984
Change-Id: Ie486ed18aab4d03b5465dbfa4201e619fef48da3