ATS: remove gerrit.wmfusercontent.org
This was added for Gerrit avatars but then we never
ended up actually enabling avatars.

The Apache site config is in sites-available but not
in sites-enabled on the backend.

Ema reported seeing some TLS errors in the logs due
to this not being in the SAN of the gerrit.wikimedia.org
cert.

Bug: T191183
Change-Id: Ie9124d5527db640201da7693099d598d226a51a5