wikidough: enable support for EDNS Client Subnet
This commit enables EDNS Client Subnet support for Wikidough by enabling
it in dnsdist and PowerDNS.

For dnsdist, it sets the useClientSubnet by adding a parameter to the
dnsdist class and sets it to true by default.

For the backend PowerDNS recursor (provided by the dnsrecusor module),
edns-subnet-whitelist is enabled by allow_edns_whitelist (default is
true) and is set to the list of WMF auth servers. Wikidough therefore
only sends the client IP address (truncated) to WMF auth servers (as
specified by $wmf_authdns in dnsrecursor). Additionally, we set
allow_incoming_ecs to true so as to make sure that PowerDNS uses the
correct source ECS information.

Bug: T252132
Change-Id: I4808db5bfeeb4d458d1d29ce5d2f631c220d3e7c