ATS: add new backend for phabricator aphlict
Remove the temporarily disabled backend on the
prod Phabricator server.

Instead add a new route to the new dedicated
backend VM aphlict1001 which is behind the new
aphlict discovery record.

envoy-proxy is installed and doing TLS termination
so it listens on 443 there while aphlict itself
is on 22280.

The "/ws/" part in the URL should not be needed anymore (!?)
as it was used to do reverse proxying to aphlict in apache.

Bug: T238593
Change-Id: I472913a95d271fabab0867f8d848cff74a3e68d6