admins: let wdqs-admins view nginx logs
The members of this admin group can restart nginx
but they can't read the error log of nginx.

journalctl -u nginx does not contain the same info
that /var/log/nginx/error.log does.

There would be other ways to do this, for example to
add group members to the "adm" group or changing permissions
on the files but this keeps it in a central location
in the admins module.

Debatable if we should stick to just nginx logs or if others
are needed as well.

Bug: T258739
Change-Id: I536f37d7ec7d8e4c510ffb50c627625f5b11c4f8