aphlict: add scap deploy target and missing parameters
In order to get the aphlict files on the new dedicated VM for
it we need to deploy Phabricator to it.

Or at least that's the easiest way forward to add the
scap::target for Phabricator as opposed to separating
the deploy repos as well.

Add the scap::target, symlink from root dir to deployment dir
and missing parameters. Copied and adjusted from the main
phabricator module.

Bug: T238593
Change-Id: Ic5603ee3ef7cab4763fa3c4c80582a78b31b3fa6