wikidough: set TLSv1.2 as the minimum version for DoT
In the current version of dnsdist's TLS configuration, the minimum
version of TLS we support is TLSv1.3, for both the DoH and DoT
frontends.

However, for DoT, one of the platforms we would like to support is
Android and its "Private DNS" mode (DoT). DoT support for Android was
added in Android 9 and support for TLSv1.3 was added in Android 10.
Given that Android 9 (39.5% of devices as of today, as per Android
Studio) has a greater market share than Android 10 (8.2% of devices),
setting the minimum version of TLS to 1.3 would break Wikidough for
Android 9 users.

For DoH, this is not an issue as Mozilla Firefox has supported TLSv1.3
for a while, even before support for DoH was added. (Google Chrome has
TLSv1.3 support since Chrome 70.)

This commit creates a separate configuration for TLS settings for DoT,
so that a minimum version of TLS (1.2) and associated ciphers can be
set, and which may differ from DoH. It also updates the dnsdist.conf.erb
template to reflect the change.

Bug: T252132
Change-Id: I7833c539cbef54a3a7bb9409fd9c0e71b72c9397